requirements.txt
# MedBlueprints Advanced AI Architecture
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Anthropic Claude API (LLM Compliance Engine)
anthropic==0.40.0

# Computer Vision
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# Graph operations (Digital Facility Graph + Regulatory Knowledge Graph)
networkx==3.3
# Neo4j driver for production graph database
neo4j==5.24.0

# Machine Learning (Approval Prediction Model)
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2

# Vector search for regulatory rules
sentence-transformers==3.1.1
faiss-cpu==1.8.0

# Data handling
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3

# Storage and serialization
sqlalchemy==2.0.35
aiosqlite==0.20.0
aiofiles==24.1.0

# PDF and CAD processing
pypdf==5.1.0
python-docx==1.1.2

# Visualization / AR output
plotly==5.24.1
shapely==2.0.6

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
loguru==0.7.2
